[cli][auth] refactor: always return CommandResult from CLI commands
Kill rich-only print-and-return-None paths so login, logout, upgrade, doctor, eval rubric, and templates share one result contract. Failed upgrade/doctor/rubric become stderr CLIError; Ctrl+C during login, upload, and log follow exits 130.
